Outline for new plugin system.
